requirements.txt
django-2.1.3
python-3
api-openweathermap.org
